+++ R/CondenseLogs.R
@@ -19,7 +19,7 @@
 #' it makes the different measures comparable. For instance, if a bone is
 #' scaled with respect to the reference, so that it homogeneously doubles its
 #' width, then all width related measures
-#' (\emph{Bd}, \emph{BT}, \emph{Bp}, and \emph{SD}) will give the
+#' (\emph{BT}, \emph{Bd}, \emph{Bp}, \emph{SD}, ...) will give the
 #' same log-ratio (\code{log(2)}). In contrast, the
 #' absolute measures are not directly comparable.
 #'
@@ -31,7 +31,7 @@
 #' mean per group, ignoring the non-available measures.
 #' We provide the following by-default group and prioritization:
 #' For lengths, the order of priority is: GL, GLl, GLm, HTC.
-#' For widths, the order of priority is: Bd, BT, Bp, SD, Bfd, Bfp.
+#' For widths, the order of priority is: BT, Bd, Bp, SD, Bfd, Bfp.
 #' This order maximises the robustness and reliability of the measurements,
 #' as priority is given to the most abundant, more replicable, and less age
 #' dependent measurements.
@@ -50,7 +50,7 @@
 #' @param grouping A list of named character vectors. The list includes a vector
 #' per selected group. Each vector gives the group of measurements in order of
 #' priority. By default the groups are \code{Length = c("GL", "GLl", "GLm", "HTC")}
-#' and \code{Width = c("Bd", "BT", "Bp", "SD", "Bfd", "Bfp")}. The order is
+#' and \code{Width = c("BT", "Bd", "Bp", "SD", "Bfd", "Bfp")}. The order is
 #' irrelevant for \code{method = "average"}.
 #' @param method Character string indicating which method to use for extracting
 #' the condensed features. Currently accepted methods: \code{"priority"}
@@ -77,7 +77,7 @@
 #'
 #' ## Extract only width with "average" method:
 #' dataExampleWithSummary2 <- CondenseLogs(dataExampleWithLogs,
-#'                                grouping = list(Width = c("Bd", "BT", "Bp", "SD")),
+#'                                grouping = list(Width = c("BT", "Bd", "Bp", "SD")),
 #'                                method = "average")
 #' head(dataExampleWithSummary2)[, -c(6:20,32:63)]