+++ NEWS.md
@@ -5,8 +5,8 @@
 2. In addition, corrected bug appearing when the method average was applied to
    a group with only one present measure log-ratio.
 3. Added some entries to the thesauri.
-4. Changed the priority order in CondenseLogs. Now "BT" has priority over "Bd" 
-   as representative of Width. 
+4. Changed the priority order in CondenseLogs following users feedback. 
+   Now "BT" has priority over "Bd" as representative of Width. 
 
 # zoolog 0.3.1
 1. Correction of error when generating vignettes with locale different from